Objects, features

The objects are the front end entities, the usual "GIS contract" objects, the features.

# S3 method for sf
sc_object(x, ...)

Arguments

x

input object

...

arguments passed to methods

See also

`sc_coord` for the coordinates part of the model, `sc_path` for the central part of the model, and `PATH` for the full model.

Examples

library(sf)
nc <-  st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE)
sc_object(nc)

#> # A tibble: 100 × 14
#>     AREA PERIMETER CNTY_ CNTY_ID        NAME   FIPS FIPSNO CRESS_ID BIR74 SID74
#>    <dbl>     <dbl> <dbl>   <dbl>      <fctr> <fctr>  <dbl>    <int> <dbl> <dbl>
#> 1  0.114     1.442  1825    1825        Ashe  37009  37009        5  1091     1
#> 2  0.061     1.231  1827    1827   Alleghany  37005  37005        3   487     0
#> 3  0.143     1.630  1828    1828       Surry  37171  37171       86  3188     5
#> 4  0.070     2.968  1831    1831   Currituck  37053  37053       27   508     1
#> 5  0.153     2.206  1832    1832 Northampton  37131  37131       66  1421     9
#> 6  0.097     1.670  1833    1833    Hertford  37091  37091       46  1452     7
#> 7  0.062     1.547  1834    1834      Camden  37029  37029       15   286     0
#> 8  0.091     1.284  1835    1835       Gates  37073  37073       37   420     0
#> 9  0.118     1.421  1836    1836      Warren  37185  37185       93   968     4
#> 10 0.124     1.428  1837    1837      Stokes  37169  37169       85  1612     1
#> # ... with 90 more rows, and 4 more variables: NWBIR74 <dbl>, BIR79 <dbl>,
#> #   SID79 <dbl>, NWBIR79 <dbl>